// src/app/interfaces/eppTorres.ts
export interface Epptorres {
    PKepptorres: number;
    fkUsuario: number;
    ArnesFPosiciones: string;
    MarcaArnes: string;
    FechaFabricacionArnes: Date;
    ComentariosArnes: string;
    BotasSeguridad: string;
    Bandola: string;
    FechaFabricacionBandola: Date;
    Guantes: string;
    LestesSeguridad: string;
    Mosqueton: string;
    ShockAbsorbente: string;
    FechaFabricacionShock: Date;
    SujetadorLineaVida: string;
    CascoBarbiquejo: string;
    Comentarios: string;
    Foto1: string;
    Foto2: string;
    FechaLlenado: Date;
    FechaActualizado: Date;
  }
  